start parsing into lines
